feat: use page helper

+++ server-i18n/server-i18n-service/src/main/java/plus/xyc/server/i18n/module/entity/response/ModuleResponse.java
@@ -24,7 +24,7 @@ public class ModuleResponse {
     @Schema(description = "目标数量")
     private Integer targetSize;
     @Schema(description = "词汇量")
-    private Integer wordSize;
+    private Long wordSize;
     @Schema(description = "更新时间")
     private java.util.Date updateTime;
     @Schema(description = "创建时间")
